requirements.txt
streamlit==1.37.1
numpy==2.3.2
Pillow==11.3.0
opencv-python==4.10.0.84
